refactor: verification module.
