+++ src/endpoints/verification/verification.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { DateTimeProxy } from '../common/date-time';
-import { RewardDto } from '../reward/dto/reward.dto';
-import { RewardService } from '../reward/reward.service';
-import { SubstrateService } from '../substrate/substrate.service';
+import { DateTimeProxy } from '../../common/date-time';
+import { RewardDto } from '../../reward/dto/reward.dto';
+import { RewardService } from '../../reward/reward.service';
+import { SubstrateService } from '../../substrate/substrate.service';
 
 @Injectable()
 export class VerificationService {
