+++ test/unit/verification/verification.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { MockType } from '../mock';
-import { VerificationService } from '../../../src/verification/verification.service';
-import { VerificationController } from '../../../src/verification/verification.controller';
+import { VerificationService } from '../../../src/endpoints/verification/verification.service';
+import { VerificationController } from '../../../src/endpoints/verification/verification.controller';
 import { ProcessEnvProxy } from '../../../src/common/process-env';
 import httpMocks = require('node-mocks-http');
